mtd-cli: Prepare for handling other endpoints
Generalise things out from being too self-assessment centric.

Split the declaration and definition of the sa_endpoint_help_map
structure array out so we have a struct endpoint_help that can be used
by other endpoints.

Generalise check_sa_args() (renaming it to check_args) so it can be used
with different endpoint_help arrays. We change its prototype somewhat to
pass in a pointer to the endpoint structure array being checked against,
the number of elements in the array and also a function pointer to the
endpoint help printer.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>