mtd-cli: Do some refactoring
Put in some infrastructure to aid in adding support for more endpoints
in the future.

It's now mainly just adjusting/creating some #defines, creating a
endpoint_help structure array, adding an entry to the new api_ep_map[]
structure array and a creating a function for calling the libmtdac API
functions.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>