endpoints: Generalise more of the endpoint infrastructure
One downside remaining with adding new endpoints is that you would have
to put an if/else block in the do_ENDPOINT() function covering each
possible endpoint.

We can get rid of this by extending the endpoint_help structure (which
has been renamed to just endpoint) to add an entry for a function
pointer to the required underlying API function.

Except that of course the API functions take differing numbers of
arguments.

However, a neat idea was found on stackoverflow[0], create a union of
function pointers, one for each function covering the number of
arguments each takes.

We can then create a function that reads through this array of endpoint
structures find the right one and call the correct function pointer,
very neat!

The mtd-cli-*.c files are really becoming very simple, but I think we
can go even further, however that's for future commits...

[0]: https://stackoverflow.com/a/16770958/11329993

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>