endpoints: Further generalisation work
This further simplifies the endpoints infrastructure.

In mtd-cli.c we rework the api_ep_map structure array, so that instead
of having a function pointer to a do_ENDPOINT() function we point to a
new structure (struct _endpoint) which contains a pointer to a struct
endpoint structure array and a function pointer to a print_endpoints()
function.

In each mtd-cli-*.c endpoint file we set this new struct _endpoint
structure and export it from the accompanying .h file.

In mtd-cli.c do_mtd_api() loops over the new api_ep_map array looking
for a match on the endpoint name, e.g 'sa' if found it then does the
check_args() call and then calls do_api_func() which will then call
right underlying API function as added in the previous commit 8dc176c
("endpoints: Generalise more of the endpoint infrastructure").

This allows us to remove the do_ENDPOINT() function entirely from each
of the mtd-cli-*.c files.

Adding new endpoints is now almost just filling out a few structures.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>