Do some adjustments to the endpoint setup and calling
This is really a preparatory patch for some future work (changing how we
pass in the optional query string arguments to the endpoint API
functions).

Rename the nargs member of the endpoint structure to nr_req_args, to
make it clear that is refers to the number of required arguments, which
may be lower than the number of arguments actually provided.

Add a new member, func, to the endpoint structure that will hold an enum
value saying how many arguments the endpoint API function takes.

The reason for separating the above two items is that nr_req_args will
be used in check_args() to check if the number of *required* arguments
has been given and func is used in do_api_func() to select the right
function pointer for the underlying API function.

The other thing we do in do_api_func() is to allocate an args[] array of
char *'s set to NULL that we populate with the values from argv[]. This
will allow for any optional query string argument (in future we will just
pass in the query string to the endpoint API functions rather than the
individual parameters), to be set to NULL.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>