mtd-cli: Re-work query string formatting
Rather than writing a function for each endpoint that takes a query
string to format it, do the actual formatting in do_api_func() then the
query string will be passed through to the underlying endpoint API
functions in the right format without needing to go through a separate
function first.

This fixes a number of endpoints which take query strings where I'd
forgot that we actually needed to make a function for them. Now they
just work.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>