mtd-cli: Allow to set additional HTTP headers
In the test environment you can often set a 'Gov-Test-Scenario' header
to affect the results you get.

libmtdac added[0] a couple of functions to allow the setting of
additional HTTP headers.

Make use of this to allow setting a 'Gov-Test-Scenario' header when
required.

This is done via a new environment variable, MTD_CLI_HDRS, e.g

    MTD_CLI_HDRS="Gov-Test-Scenario: NO_SUBMISSIONS_EXIST" mtd-cli sa cr-intent-to-crystallise 2018-19

Currently mtd-cli only supports the setting of a single header (libmtdac
how no such limitation) which should be enough for this use case.

[0]: https://github.com/ac000/libmtdac/commit/e2947c2837783b76ca70f4487f344acde1b1e50b

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>