Update for change in POST/PUT endpoints API
With commit 08eb2cf ("Totally re-work how the source JSON payload is
specified") of libmtdac[0] the way that the JSON payload data is
specified has changed. Rather than just passing in a filename we now
pass in a struct mtd_dsrc_ctx that describes how the JSON data is being
provided, it can now be either a filename, buffer, stdio FILE * file
pointer or a file descriptor.

The main change this required was to create new function pointers as now
some functions take a const char * as their first argument and some take
a const struct mtd_dsrc_ctx *

We also create some new matching FUNC_ enums so the right function is
selected.

[0]: https://github.com/ac000/libmtdac/commit/08eb2cfbada82745c498041bb79ae1c2871e09c4

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>