mtd-cli.h: Allow to auto-generate the endpoint structure name
Add some macro magic[0] to allow to auto-generate the endpoint structure
name. E.g this allows us to do (in the mtd-cli-*.c files)

    #define API        ni
    ...
    const struct _endpoint ENDPOINT = {

which gets turned into

    const struct _endpoint ni_endpoint = {

The tricky bit is we need to force the CPP to do two passes over the
expansion, as simply doing something like

    #define API        ni
    #define ENDPOINT   API ##_endpoint

just results in

    const struct _endpoint API_endpoint = {

The next commit will convert the various mtd-cli-*.c files over to this.

[0]: http://jhnet.co.uk/articles/cpp_magic (Forcing CPP to make multiple
passes)

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>