Auto-generate the struct _endpoint name in mtd-cli-*.c
Do some cleanup in the mtd-cli-*.c files by making use of some
infrastructure added in the previous commit 56b90fd ("mtd-cli.h: Allow
to auto-generate the endpoint structure name") by defining the API name,
e.g

    #define API    ni

this is then used to name the struct _endpoint as such

    const struct _endpoint ENDPOINT = {

which in this example becomes

    const struct _endpoint ni_endpoint = {

So we now have one less thing to change if we are using an existing file
as a template for a new API endpoint.

This also gets rid of the need for the MTD_CLI_CMD define with some
small change to mtd-cli.c. So another nice cleanup as we now don't need
to specify that in every '.use'.

Technically we could take it further by skipping the API endpoint cmd
name from .use as we already have it in .name...

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>