Remove the use of jansson
Beginning with commit 609b24a ("curler: Fix formatting of real numbers
in the JSON response") of libmtdac[0], libmtdac now formats the JSON in
a sensible way, it indents the JSON with four spaces and tries to have
real numbers displayed properly to two decimal places.

With the above there is no need now to process the JSON so just remove
the jansson stuff and print the result buffer as is.

[0]: https://github.com/ac000/libmtdac/commit/609b24adaaa355a8ffa50e0bafd3a1c07938790d

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>