Remove some redundancy from the endpoint.use member
In struct endpoint the 'use' member contained the full command name as
well as any arguments it took, e.g

    "get businessId"

however we already have the command name in the 'name' member.

Lets remove the command name from 'use' (rename it to 'args' while were
at it for a more appropriate name now). So now the above is simply

    "businessId"

when printing out the use/help text we now print the 'name' member
before 'use' so everything is as it was.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>