Big update for *new* libmtdac
libmtdac has undergone a major revamp, this aligns to it.

Various endpoint commands have been removed (as they have in libmtdac).
It's likely some (and others) will be added back as and when...

Most of the endpoint commands documentation has been removed as it's
outdated and it really just duplicates the official documentation. Also
the commands somewhat document what they expect. E.g.

  $ ./mtd-cli API endpoint ?

will show what arguments that command expects.

Signed-off-by: Andrew Clayton <ac@sigsegv.uk>